// src/types/database.ts
export type IncidentSeverity = "CRITICAL" | "HIGH" | "MEDIUM" | "LOW";
export type IncidentType =
  | "FLOOD"
  | "FIRE"
  | "LANDSLIDE"
  | "STRUCTURAL_COLLAPSE"
  | "MEDICAL_EMERGENCY"
  | "EARTHQUAKE"
  | "CYCLONE"
  | "OTHER";